downloads/requirements.txt
Django
django-widget-tweaks
python-dotenv==1.0.1
python-ulid==1.1.0
requests==2.32.3
django-filter==24.2
python-dotenv